LayoutTests/canvas/philip/tests/2d.path.isPointInPath.transform.3-expected.txt

< [index] >
2d.path.isPointInPath.transform.3
isPointInPath() handles transformations correctly
References: 2d.path.isPointInPath Bugzilla
Actual output:
Passed
